+++ umpa-core-api/src/main/kotlin/com/umpa/core/domain/content/playlist/PlaylistCreateService.kt
@@ -7,9 +7,7 @@ import org.springframework.web.multipart.MultipartFile
 @Service
 class PlaylistCreateService(
     private val playlistCreator: PlaylistCreatorWrapper,
-    private val userReader: UserReader,
-    private val playlistUpdater: PlaylistUpdaterWrapper,
-    private val playlistDetailFiller: PlaylistDetailFiller
+    private val userReader: UserReader
 ) {
     fun create(creation: PlaylistCreation): PlaylistDetail {
         val playlist = playlistCreator.create(creation)
@@ -31,10 +29,5 @@ class PlaylistCreateService(
 
     private fun imageUploadAndUpdate(image: MultipartFile): String {
         return "url"
-    }
-
-    fun edit(revision: PlaylistRevision): PlaylistDetail {
-        val playlist = playlistUpdater.edit(revision)
-        return playlistDetailFiller.fill(playlist)
     }
 }
